+++ src/graphql/account.rs
@@ -194,6 +194,7 @@ impl AccountMutation {
                     &None,
                     &None,
                     &None,
+                    &None,
                 )?;
                 return Ok(username);
             }
@@ -235,6 +236,7 @@ impl AccountMutation {
         role: Option<UpdateRole>,
         name: Option<UpdateName>,
         department: Option<UpdateDepartment>,
+        language: Option<UpdateLanguage>,
         allow_access_from: Option<UpdateAllowAccessFrom>,
         max_parallel_sessions: Option<UpdateMaxParallelSessions>,
     ) -> Result<String> {
@@ -242,6 +244,7 @@ impl AccountMutation {
             && role.is_none()
             && name.is_none()
             && department.is_none()
+            && language.is_none()
             && allow_access_from.is_none()
             && max_parallel_sessions.is_none()
         {
@@ -251,6 +254,7 @@ impl AccountMutation {
         let role = role.map(|r| (database::Role::from(r.old), database::Role::from(r.new)));
         let name = name.map(|n| (n.old, n.new));
         let dept = department.map(|d| (d.old, d.new));
+        let language = language.map(|d| (d.old, d.new));
         let allow_access_from = if let Some(ip_addrs) = allow_access_from {
             let old = if let Some(old) = ip_addrs.old {
                 Some(strings_to_ip_addrs(&old)?)
@@ -276,6 +280,7 @@ impl AccountMutation {
             role,
             &name,
             &dept,
+            &language,
             &allow_access_from,
             &max_parallel_sessions,
         )?;
@@ -467,6 +472,10 @@ impl Account {
         &self.inner.department
     }
 
+    async fn language(&self) -> Option<String> {
+        self.inner.language.clone()
+    }
+
     async fn creation_time(&self) -> DateTime<Utc> {
         self.inner.creation_time()
     }
@@ -536,6 +545,12 @@ struct UpdateName {
 struct UpdateDepartment {
     old: String,
     new: String,
+}
+
+#[derive(InputObject)]
+struct UpdateLanguage {
+    old: Option<String>,
+    new: Option<String>,
 }
 
 /// The old and new values of `allowAccessFrom` to update.
@@ -1121,7 +1136,8 @@ mod tests {
                         password: "Ahh9booH",
                         role: "SECURITY_ADMINISTRATOR",
                         name: "John Doe",
-                        department: "Security"
+                        department: "Security",
+                        language: "en-US"
                     )
                 }"#,
             )
@@ -1136,7 +1152,8 @@ mod tests {
                         password: "Ahh9booH",
                         role: "SYSTEM_ADMINISTRATOR",
                         name: "John Doe",
-                        department: "Admin"
+                        department: "Admin",
+                        language: "en-US"
                     )
                 }"#,
             )
@@ -1182,6 +1199,98 @@ mod tests {
             )
             .await;
         assert_eq!(res.data.to_string(), r#"null"#);
+    }
+
+    #[tokio::test]
+    async fn update_account() {
+        let schema = TestSchema::new().await;
+
+        let res = schema
+            .execute(
+                r#"mutation {
+                    insertAccount(
+                        username: "username",
+                        password: "password",
+                        role: "SECURITY_ADMINISTRATOR",
+                        name: "John Doe",
+                        department: "Security",
+                        language: "en-US"
+                    )
+                }"#,
+            )
+            .await;
+
+        assert_eq!(res.data.to_string(), r#"{insertAccount: "username"}"#);
+
+        let res = schema
+            .execute(
+                r#"
+                query {
+                     account(username: "username") {
+                        username
+                        role
+                        name
+                        department
+                        language
+                    }
+                }"#,
+            )
+            .await;
+
+        assert_eq!(
+            res.data.to_string(),
+            r#"{account: {username: "username", role: SECURITY_ADMINISTRATOR, name: "John Doe", department: "Security", language: "en-US"}}"#
+        );
+
+        let res = schema
+            .execute(
+                r#"
+                mutation {
+                    updateAccount(
+                        username: "username",
+                        password: "password",
+                        role: {
+                            old: "SECURITY_ADMINISTRATOR",
+                            new: "SYSTEM_ADMINISTRATOR"
+                        },
+                        name: {
+                            old: "John Doe",
+                            new: "Loren Ipsum"
+                        },
+                        department: {
+                            old: "Security",
+                            new: "Admin"
+                        },
+                        language: {
+                            old: "en-US",
+                            new: "ko-KR"
+                        }
+                    )
+                }"#,
+            )
+            .await;
+
+        assert_eq!(res.data.to_string(), r#"{updateAccount: "username"}"#);
+
+        let res = schema
+            .execute(
+                r#"
+                query {
+                     account(username: "username") {
+                        username
+                        role
+                        name
+                        department
+                        language
+                    }
+                }"#,
+            )
+            .await;
+
+        assert_eq!(
+            res.data.to_string(),
+            r#"{account: {username: "username", role: SYSTEM_ADMINISTRATOR, name: "Loren Ipsum", department: "Admin", language: "ko-KR"}}"#
+        );
     }
 
     #[tokio::test]